python-example-app/requirements.txt
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-sdk==1.20.0
Flask
Flask_RESTful
neo4j
scikit_learn
